move writer methods to new writer class (#33)
