+++ src/test/java/com/imsweb/x12/reader/X12ReaderTest.java
@@ -25,8 +25,10 @@ import org.junit.Test;
 import com.imsweb.x12.Element;
 import com.imsweb.x12.LineBreak;
 import com.imsweb.x12.Loop;
+import com.imsweb.x12.X12;
 import com.imsweb.x12.mapping.TransactionDefinition;
 import com.imsweb.x12.reader.X12Reader.FileType;
+import com.imsweb.x12.writer.X12Writer;
 
 public class X12ReaderTest {
 
@@ -51,82 +53,7 @@ public class X12ReaderTest {
         assertEquals(fromFileUtf8.getLoops().get(0).toString(), fromInputStreamUtf8.getLoops().get(0).toString());
         assertEquals(fromFileUtf8.getLoops().get(0).toString(), fromReaderUtf8.getLoops().get(0).toString());
     }
-
-    /**
-     * Here we will test that you can go from x12, make changes, then serialize the
-     * x12 once again.
-     */
-    @Test
-    public void testSerializeBasic() throws IOException {
-        URL url = this.getClass().getResource("/837_5010/x12_valid.txt");
-
-        X12Reader fromFileUtf8 = new X12Reader(FileType.ANSI837_5010_X222, new File(url.getFile()),
-                StandardCharsets.UTF_8);
-
-        String expected = IOUtils
-                .toString(this.getClass().getResourceAsStream("/837_5010/x12_valid.txt"), StandardCharsets.UTF_8)
-                .trim();
-        LineBreak lineBreak;
-        if (expected.contains(LineBreak.CRLF.getLineBreakString())) {
-            lineBreak = LineBreak.CRLF;
-        }
-        else {
-            lineBreak = LineBreak.LF;
-        }
-        Assert.assertEquals(expected, fromFileUtf8.toX12String(lineBreak).trim());
-    }
-
-    /**
-     * Tests the toHtml method that
-     */
-    @Test
-    public void testToHtmlBasic() throws IOException {
-        URL url = this.getClass().getResource("/837_5010/x12_valid.txt");
-
-        X12Reader fromFileUtf8 = new X12Reader(FileType.ANSI837_5010_X222, new File(url.getFile()),
-            StandardCharsets.UTF_8);
-
-        String x12Template = IOUtils.toString(getClass().getResourceAsStream("/html/x12-template.html"), StandardCharsets.UTF_8);
-
-        String x12HtmlSegment = fromFileUtf8.toHtml();
-
-        String fullX12Html = String.format(x12Template, x12HtmlSegment);
-
-        Document doc = Jsoup.parse(fullX12Html);
-        Elements loops = doc.select(".x12-loop");
-        Assert.assertEquals(20, loops.size());
-
-        Elements segments = doc.select(".x12-segment");
-        Assert.assertEquals(38, segments.size());
-
-        Elements elements = doc.select(".x12-element");
-        Assert.assertEquals(216, elements.size());
-    }
-
-    /**
-     * Test a more complex x12 doc and see if we can serialize it.
-     */
-    @Test
-    public void testSerializeComplex() throws IOException {
-        URL url = this.getClass().getResource("/837_5010/x12_complex.txt");
-
-        X12Reader fromFileUtf8 = new X12Reader(FileType.ANSI837_5010_X222, new File(url.getFile()),
-                StandardCharsets.UTF_8);
-
-        String expected = IOUtils
-                .toString(this.getClass().getResourceAsStream("/837_5010/x12_complex.txt"), StandardCharsets.UTF_8)
-                .trim();
-
-        LineBreak lineBreak;
-        if (expected.contains(LineBreak.CRLF.getLineBreakString())) {
-            lineBreak = LineBreak.CRLF;
-        }
-        else {
-            lineBreak = LineBreak.LF;
-        }
-        Assert.assertEquals(expected, fromFileUtf8.toX12String(lineBreak).trim());
-    }
-
+    
     @Test
     public void testMultipleGSLoops() throws Exception {
         URL url = this.getClass().getResource("/837_5010/x12_multiple_gs.txt");
